+++ cartridge/confapplier.lua
@@ -38,6 +38,7 @@ local OperationError = errors.new_class('OperationError')
 
 vars:new('state', '')
 vars:new('error')
+vars:new('timestamp') -- used for issues
 vars:new('state_notification', fiber.cond())
 vars:new('state_notification_timeout', 5)
 vars:new('clusterwide_config')
@@ -140,6 +141,7 @@ local function set_state(new_state, err)
     membership.set_payload('state', new_state)
     vars.state = new_state
     vars.error = err
+    vars.timestamp = fiber.time()
     vars.state_notification:broadcast()
 end